Generate only a single manual.
Added the ability to pass the .md document template as argument to manning-up.sh which will only generate the single document and rename the .1.html correctly before exiting and thus not building the index and all the docs again.
Includes updated documentation.

+++ man/manning-up.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+test "$1" && ronn $1 && mv -f ${1/.md/}.1.html ${1/.md/}.html && exit 0
+
 echo '# manuals' > index.txt.tmp
 ln=$(awk '/## COMMANDS/{print NR};'  ./git-extras.md)
 awk "NR <= $ln+1" git-extras.md > git-extras.md.tmp
